backend/pyproject.toml
[tool.poetry]
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
aiosqlite = "^0.20.0"
alchemical = "^1.0.2"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
argon2-cffi = "^23.1.0"
fastapi-mail = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
